style 3단계 - 생성자 개행 스타일 변경
- 생성자 단일 통합
- integer -> int 타입 수정

+++ src/main/java/gameboy/gamepack/racinggame/data/vo/Position.java
@@ -11,8 +11,8 @@ public class Position {
         this.position = position;
     }
 
-    private void validationPosition(Integer position) {
-        if (position.intValue() <= 0) {
+    private void validationPosition(int position) {
+        if (position <= 0) {
             throw new IllegalArgumentException("0이하의 포지션 생성: " + position);
         }
     }
